requirements-dev.txt
apache-airflow==2.9.2
pytrends==4.9.2
yfinance==1.2.1
pandas==2.2.2
psycopg2-binary==2.9.9
streamlit
plotly
jupyter